+++ tests/test_associated_future.rs
@@ -17,7 +17,7 @@ mod without_unmock {
 
     #[tokio::test]
     async fn should_mock_without_unmock() {
-        let unimock = mock(FooMock::without_unmock.each_call(matching!(10)).returns(20));
+        let unimock = Unimock::new(FooMock::without_unmock.each_call(matching!(10)).returns(20));
 
         let answer = unimock.without_unmock(10).await;
 
@@ -43,9 +43,9 @@ mod with_unmock {
 
     #[tokio::test]
     async fn should_mock_with_unmock() {
-        let unimock = mock(FooMock::with_unmock.each_call(matching!(42)).unmocked());
+        let test = Unimock::new(FooMock::with_unmock.each_call(matching!(42)).unmocked());
 
-        let answer = unimock.with_unmock(42).await;
+        let answer = test.with_unmock(42).await;
 
         assert_eq!(84, answer);
     }